pyproject.toml
[project]
name = "anki-maxxing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.20.0",
    "anthropic>=0.40.0",
    "beautifulsoup4>=4.12.0",
    "cryptography>=43.0.0",
    "fastapi>=0.115.0",
    "google-genai>=1.0.0",
    "greenlet>=3.4.0",
    "httpx>=0.27.0",
    "lxml>=5.3.0",
    "openai>=1.50.0",
    "pydantic>=2.8.0",
    "pydantic-settings>=2.4.0",
    "python-dateutil>=2.9.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.12",
    "sqlalchemy>=2.0.0",
    "structlog>=24.4.0",
    "tenacity>=9.0.0",
    "uvicorn[standard]>=0.30.0",
]
